% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bpcaNet}
\alias{bpcaNet}
\title{Bayesian PCA updates}
\usage{
bpcaNet(
  myMat,
  covy,
  N,
  D,
  hidden,
  numberOfNonNAvaluesInEachCol,
  nomissIndex,
  missIndex,
  nMissing,
  nPcs = 2L,
  threshold = 1e-04,
  maxIterations = 200L
)
}
\arguments{
\item{myMat}{\code{matrix} -- data matrix with observations in rows and 
variables in columns. (Note that this is the transpose of \code{X} in
\code{\link{pca_full}}.)}

\item{covy}{\code{matrix} -- the unbiased sample covariance of the data matrix.}

\item{N}{\code{numeric} -- the number of observations.}

\item{D}{\code{numeric} -- the number of variables.}

\item{hidden}{\code{numeric} -- indices of missing values in \code{1:length(myMat)}.}

\item{numberOfNonNAvaluesInEachCol}{\code{numeric} -- number of observed values in each 
column of the data (i.e. variables).}

\item{nomissIndex}{\code{numeric} -- indices of rows (observations) without any
missing values.}

\item{missIndex}{\code{numeric} -- indices of rows (observations) with missing
values.}

\item{nMissing}{\code{numeric} -- total number of missing values.}

\item{nPcs}{\code{numeric} -- number of components/latent variables to use.}

\item{threshold}{\code{numeric} -- threshold for convergence, applied to the precision
parameter \code{tau}. Updates for which the change in \code{tau} are below this threshold
value stop the algorithm.}

\item{maxIterations}{\code{numeric} -- the maximum number of iterations to be completed.}
}
\value{
{A \code{list} of 5 elements:
\describe{
\item{W}{\code{matrix} -- the estimated loadings.}
\item{ss}{\code{numeric} -- the estimated model variance.}
\item{C}{\code{matrix} -- the estimated covariance matrix.}
\item{scores}{\code{matrix} -- the estimated scores.}
\item{m}{\code{numeric} -- the estimated mean vector.}
}}
}
\description{
Perform parameter updates for PPCA using the Variational Bayes framework
from Oba (2003). Not recommended to use standalone, rather it is called from within
\code{\link{bpcapM}} and its wrapper \code{\link{pcapM}}.
}
\examples{
set.seed(102)
  N <- 20
  D <- 20
  nPcs <- 2
  maxIterations <- 1000
  X <- matrix(rnorm(50), D, N)
  X <- scale(X, center=TRUE, scale=FALSE) # mean 0
  covX <- cov(X)
  IX <- sample(1:D, 10)
  JX <- sample(1:N, 10)
  nMissing <- length(IX)+length(JX)
  X[JX, IX] <- 0
  hidden <- which(X==0)
  numberOfNonNAvaluesInEachCol <- colSums(X!=0)
  nomissIndex <- which(rowSums(X!=0)==N)
  missIndex <- which(rowSums(X!=0)!=N)
  threshold <- 1e-4
  bpcaNetOutput <- bpcaNet(myMat=X, covy=covX, N=N, D=D, hidden=hidden,
    numberOfNonNAvaluesInEachCol=numberOfNonNAvaluesInEachCol,
    nomissIndex=nomissIndex, missIndex=missIndex, nMissing=nMissing,
    nPcs=nPcs, threshold=threshold, maxIterations=maxIterations)
}
\references{
Oba, S., Sato, M.A., Takemasa, I., Monden, M.,
 Matsubara, K.I. and Ishii, S., 2003.
 \href{https://doi.org/10.1093/bioinformatics/btg287}{doi}.
 
 Stacklies, W., Redestig, H., Scholz, M., Walther, D. and 
 Selbig, J., 2007.
 \href{https://doi.org/10.1093/bioinformatics/btm069}{doi}.
}
\seealso{
\code{\link{bpcapM}}, \code{\link{pcapM}}
}
