% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_full.R
\name{compute_rms}
\alias{compute_rms}
\title{Compute the root mean-squared error of a PCA projection}
\usage{
compute_rms(X, A, S, M, ndata, verbose = TRUE)
}
\arguments{
\item{X}{\code{matrix} -- the data matrix with variables in rows and 
observations in columns.}

\item{A}{\code{matrix} -- initialised loadings matrix with observed variables
in rows and latent variables in columns.}

\item{S}{\code{matrix} -- initialised factor scores matrix with latent variables
in rows and observations in columns.}

\item{M}{\code{matrix} -- logical matrix whose values indicate whether
the corresponding entry in \code{X} is observed.}

\item{ndata}{\code{numerical} -- the total number of observed values.}

\item{verbose}{\code{logical} -- whether extra output should be displayed.}
}
\value{
{A \code{list} of length 2:
 \describe{
 \item{errMx}{\code{matrix} -- matrix of element-wise differences (errors)
 between the observed data and the PCA projection.}
 \item{rms}{\code{numerical} -- root mean-squared error of the PCA
 projection.}
 }
}
}
\description{
Root mean-squared error is the square root of the element-wise error's mean.
 This is a useful quantity to display during parameter estimation in \code{\link{pca_updates}}
 since it is a measure of how well the PCA projection is fitting the data.
}
\examples{
p <- 20
n <- 7
set.seed(10045)
X <- matrix(rnorm(p*n), p, n)
miss.inds <- sample(1:(p*n), (p*n)/4)
X[miss.inds] <- NA
M <- !is.na(X)
Nobs_i <- rowSums(M)
Mu <- rowSums(X, na.rm = TRUE) / Nobs_i
update_bias <- TRUE
Xcent <- subtractMu(Mu=Mu, X=X, M=M, p=p, n=n, update_bias=update_bias, verbose=TRUE)
init.model <- initParms(p=p, n=n, ncomp=2, verbose = TRUE)
compute_rms(X=X, A=init.model$A, S=init.model$S, M=M, ndata=sum(Nobs_i), verbose=TRUE)
}
