% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_full.R
\name{initParms}
\alias{initParms}
\title{Initialise model parameters for \code{\link{pca_updates}}}
\usage{
initParms(p, n, ncomp, verbose = TRUE)
}
\arguments{
\item{p}{\code{numeric} -- the number of variables}

\item{n}{\code{numeric} -- the number of observations}

\item{ncomp}{\code{numeric} -- the number of components/latent variables}

\item{verbose}{\code{logical} -- whether extra output should be displayed}
}
\value{
{A \code{list} of length 7:
 \describe{
 \item{A}{\code{matrix} -- initialised loadings matrix with observed variables
  in rows and latent variables in columns.}
 \item{S}{\code{matrix} -- initialised factor scores matrix with latent variables
  in rows and observations in columns.}
 \item{V}{\code{numeric} -- scalar value corresponding to the initialised 
 variance of the error parameter.}
 \item{Av}{\code{array} -- initialised covariance matrices of the rows of A.}
 \item{Sv}{\code{array} -- initialised covariance matrices of the rows of S.}
 \item{Muv}{\code{numeric} -- the initialisation of the prior variance of Mu.}
 }
}
}
\description{
Internal function within \code{\link{pca_full}} that initialises
 most model parameters. WARNING: does not initialise all parameters by itself
 correctly (since this depends on context) and so care should be taken when 
 using as a standalone function.
}
\details{
Random initialisations are set for the loadings and scores matrices.
 Diagonal matrices are set for the elements of \code{Av} and \code{Sv}. \code{V}
 is initialised to 1 and \code{Muv} is initialised to a vector of 1s.
}
\examples{
init.model <- initParms(p=10, n=10, ncomp=2, verbose = TRUE)
init.model$A
init.model$Av
}
