\name{pcaNet-package}
\alias{pcaNet-package}
\alias{pcaNet}
\docType{package}
\title{
\packageTitle{pcaNet}
}
\description{
\packageDescription{pcaNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pcaNet}
\packageIndices{pcaNet}
An overview of how to use the package, including the most important
functions
}
\author{
\packageAuthor{pcaNet}

Maintainers: \packageMaintainer{pcaNet}
}
\references{
Oba, S., Sato, M.A., Takemasa, I., Monden, M.,
Matsubara, K.I. and Ishii, S., 2003. \href{https://doi.org/10.1093/bioinformatics/btg287}{doi}.

Stacklies, W., Redestig, H., Scholz, M., Walther, D. and 
Selbig, J., 2007. \href{https://doi.org/10.1093/bioinformatics/btm069}{doi}.
}
\keyword{ package }
\seealso{
\code{\link[pcaMethods:pcaMethods-package]{pcaMethods}}
}