% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppca2Covinv.R
\name{ppca2Covinv}
\alias{ppca2Covinv}
\title{Inverse covariance matrix computation from PPCA}
\usage{
ppca2Covinv(ppcaOutput)
}
\arguments{
\item{ppcaOutput}{\code{list} -- the output object from running any
of the PPCA functions in this package.}
}
\value{
\code{matrix} -- the inverse of the covariance matrix.
}
\description{
Efficient inversion of the covariance matrix
  estimated from PPCA.
}
\details{
The computation exploits the
  Woodbury identity so that a kxk matrix (where k is often less than 10)
  is inverted instead of the potentially large pxp matrix. The closed-form
  expression for the inverse depends upon parameters that are estimated
  in the PPCA algorithm.
}
