\name{ppcaNet-package}
\alias{ppcaNet-package}
\alias{ppcaNet}
\docType{package}
\title{
\packageTitle{ppcaNet}
}
\description{
\packageDescription{ppcaNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ppcaNet}
\packageIndices{ppcaNet}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ppcaNet}

Maintainer: \packageMaintainer{ppcaNet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
