% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ppcaNet}
\alias{ppcaNet}
\title{Probabilistic PCA updates}
\usage{
ppcaNet(myMat, N, D, W, hidden, nMissing, nPcs = 2L, threshold = 1e-05,
  maxIterations = 1000L)
}
\arguments{
\item{myMat}{\code{matrix} -- data matrix with observations in rows and 
variables in columns. (Note that this is the transpose of \code{X} in
\code{\link{pca_full}}.)}

\item{N}{\code{numeric} -- the number of observations.}

\item{D}{\code{numeric} -- the number of variables.}

\item{W}{\code{matrix} -- initialised loadings matrix with observed
variables in rows and latent variables in columns.}

\item{hidden}{\code{numeric} -- indices of missing values in \code{1:length(myMat)}.}

\item{nMissing}{\code{numeric} -- total number of missing values.}

\item{nPcs}{\code{numeric} -- number of components/latent variables to use.}

\item{threshold}{\code{numeric} -- threshold for convergence, applied to the precision
parameter \code{tau}. Updates for which the change in \code{tau} are below this threshold
value stop the algorithm.}

\item{maxIterations}{\code{numeric} -- the maximum number of iterations to be completed.}
}
\value{
{A \code{list} of 4 elements:
\describe{
\item{W}{\code{matrix} -- the estimated loadings.}
\item{ss}{\code{numeric} -- the estimated model variance.}
\item{C}{\code{matrix} -- the estimated covariance matrix.}
\item{myMat}{\code{matrix} -- the data matrix with missing values
replaced by their estimated projections.}
}}
}
\description{
Perform parameter updates for PPCA using the Expectation-Maximisation framework
from Porta (2005) and also in the R-package \code{\link{pcaMethods}} (Stacklies, 2007).
Not recommended to use standalone, rather it is called from within
\code{\link{ppcapM}} and its wrapper \code{\link{pcapM}}.
}
\examples{
set.seed(102)
N <- 20
D <- 20
nPcs <- 2
maxIterations <- 1000
X <- matrix(rnorm(50), D, N)
X <- scale(X, center=TRUE, scale=FALSE) # mean 0
covX <- cov(X)
IX <- sample(1:D, 10)
JX <- sample(1:N, 10)
nMissing <- length(IX)+length(JX)
X[JX, IX] <- 0
hidden <- which(X==0)
threshold <- 1e-4
r <- sample(N)
W <- t(X[r[1:nPcs], ,drop = FALSE])
W <- matrix(rnorm(W), nrow(W), ncol(W), dimnames = labels(W) )
ppcaNetOutput <- ppcaNet(X, N, D, W, hidden, nMissing, nPcs, threshold, maxIterations)

}
\references{
Porta, J.M., Verbeek, J.J. and Kroese, B.J., 2005.
 \href{https://hal.inria.fr/inria-00321476/en}{link}
 
 Stacklies, W., Redestig, H., Scholz, M., Walther, D. and 
 Selbig, J., 2007. \href{https://doi.org/10.1093/bioinformatics/btm069}{doi}.
}
\seealso{
\code{\link{ppcapM}}, \code{\link{pcapM}}
}
