% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcapM.R
\name{ppcapM}
\alias{ppcapM}
\title{Title}
\usage{
ppcapM(myMat, nPcs = 2, seed = NA, threshold = 1e-05,
  maxIterations = 1000, ...)
}
\arguments{
\item{myMat}{}

\item{nPcs}{}

\item{seed}{}

\item{threshold}{}

\item{maxIterations}{}

\item{...}{}
}
\description{
Title
}
