% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spcapp}
\alias{spcapp}
\title{Title}
\usage{
spcapp(myMat, nPcs, maxIterations = 10000, TolFun = 1e-06, TolX = 1e-06)
}
\arguments{
\item{myMat}{}

\item{nPcs}{}

\item{maxIterations}{}

\item{TolFun}{}

\item{TolX}{}
}
\description{
Title
}
