% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_full.R
\name{subtractMu}
\alias{subtractMu}
\title{Subtract the row means from a matrix of data with missing values}
\usage{
subtractMu(Mu, X, M, p, n, update_bias, verbose = TRUE)
}
\arguments{
\item{Mu}{\code{numeric} -- the sample mean of the observed variables.}

\item{X}{\code{matrix} -- the data matrix with variables in rows and 
observations in columns.}

\item{M}{\code{matrix} -- logical matrix whose values indicate whether
the corresponding entry in \code{X} is observed.}

\item{p}{\code{numeric} -- the number of variables.}

\item{n}{\code{numeric} -- the number of observations.}

\item{update_bias}{\code{logical} -- whether the mean should be subtracted.
or not.}

\item{verbose}{\code{logical} -- whether extra output should be displayed.}
}
\value{
\code{X} \code{matrix} -- centered data matrix.
}
\description{
internal function within \code{\link{pca_full}} to subtract the row means
 from a matrix of data using only the observed values. Offers little utility 
 standalone.
}
\examples{
p <- 20
n <- 7
set.seed(10045)
X <- matrix(rnorm(p*n), p, n)
miss.inds <- sample(1:(p*n), (p*n)/4)
X[miss.inds] <- NA
M <- !is.na(X)
Nobs_i <- rowSums(M)
Mu <- rowSums(X, na.rm = TRUE) / Nobs_i
update_bias <- TRUE
Xcent <- subtractMu(Mu=Mu, X=X, M=M, p=p, n=n, update_bias=update_bias, verbose=TRUE)
X-Xcent
Mu # all observed values in each column equal to Mu
}
