% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbpca.R
\name{vbpca}
\alias{vbpca}
\title{Title}
\usage{
vbpca(myMat, nPcs = NA, maxIterations = 10000, TolFun = 1e-04,
  TolX = 1e-04)
}
\arguments{
\item{myMat}{}

\item{nPcs}{}

\item{maxIterations}{}

\item{TolFun}{}

\item{TolX}{}
}
\description{
Title
}
