% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbpca.R
\name{vbpcapp}
\alias{vbpcapp}
\title{Title}
\usage{
vbpcapp(myMat, nPcs = 2, maxIterations = 10000, TolFun = 1e-04,
  TolX = 1e-04)
}
\arguments{
\item{myMat}{}

\item{nPcs}{}

\item{maxIterations}{}

\item{TolFun}{}

\item{TolX}{}
}
\description{
Title
}
