% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpcapM.R
\name{bpcapM}
\alias{bpcapM}
\title{Bayesian PCA missing value estimation (pcaMethods version)}
\usage{
bpcapM(myMat, nPcs = 2, threshold = 1e-04, maxIterations = 100, ...)
}
\arguments{
\item{myMat}{\code{matrix} -- Pre-processed matrix (centered,
scaled) with variables in columns and observations in rows. The
data may contain missing values, denoted as \code{NA}.}

\item{nPcs}{\code{numeric} -- Number of components used for
re-estimation. Choosing few components may decrease the
estimation precision.}

\item{threshold}{convergence threshold.}

\item{maxIterations}{\code{numeric} -- Maximum number of estimation
steps.}

\item{...}{}
}
\description{
Implements a Bayesian PCA missing value estimator, as in pcaMethods.
}
\details{
Details about the probabilistic model underlying BPCA are found in
Oba et. al 2003. The algorithm uses an expectation maximation
approach together with a Bayesian model to approximate the
principal axes (eigenvectors of the covariance matrix in PCA).
The estimation is done iteratively, the algorithm terminates if
either the maximum number of iterations was reached or if the
estimated increase in precision falls below \eqn{1e^{-4}}{1e^-4}.
}
