\name{pcaNet-package}
\alias{pcaNet-package}
\alias{pcaNet}
\docType{package}
\title{
\packageTitle{pcaNet}
}
\description{
\packageDescription{pcaNet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pcaNet}
\packageIndices{pcaNet}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pcaNet}

Maintainer: \packageMaintainer{pcaNet}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}