% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppca2Covplot.R
\name{ppca2Covplot}
\alias{ppca2Covplot}
\title{Covariance matrix visualisation}
\usage{
ppca2Covplot(ppcaOutput)
}
\arguments{
\item{ppcaOutput}{\code{list} -- the output object from running any
of the PPCA functions in this package.}
}
\description{
Heatmap visualisation of the covariance matrix
   estimated within PPCA.
}
